Move CloudRegion and REGION_LABELS to shared/types/regions
