--- apps/code/src/shared/types/oauth.ts
@@ -1,1 +1,0 @@
-export type CloudRegion = "us" | "eu" | "dev";
